test: replaced LoginScreenTest ui tests with mockk & robolectric
